Create Median of two sorted arrays.cpp
Time complexity: o(nlogn)
Space Complexity: o(n)...

The time complexity of the given code is O((n+m)log(n+m)), where n and m are the sizes of the input vectors n1 and n2, respectively.

The reasoning behind this is as follows:

Initializing the ans vector with the elements of n1 and n2 takes O(n+m) time.
Sorting the ans vector takes O((n+m)log(n+m)) time, due to the use of the sort() function.
Calculating the median based on the sorted ans vector takes constant time, O(1).
Therefore, the overall time complexity is O((n+m)log(n+m)).

The space complexity of the given code is O(n+m), as we are creating an additional vector ans to hold the combined elements of n1 and n2. In addition, we are also creating vectors n1 and n2 of size n and m, respectively. However, as these are input vectors, they are not part of the space complexity of the algorithm itself.

Therefore, the overall space complexity is O(n+m).